fix bug in partition id generation
there was a bug in the generation of the partition ids. only one single
instance of the hasher has been used, every time a partition id needs to
be generated the hasher first gets reset, then the data to be hashed
gets written into it, then the hash is generated. since tsdb-gw is multi
threaded, multiple threads could have used that same hasher at the same
time.