regularly force-refresh keycache to ensure periodic sending of MetricData
Metrictank users expect to be able to remove entries from the index,
and to have them re-appear if they keep sending data.
A simple way to achieve this is for tsdb-gw to regularly resend full
MetricData payloads, just in case users have deleted important index
entries.
Obviously this comes with a performance trade-off but the overhead
should be minimal.