refactor token and instance caches. fixes #130
Updating entries in the cache has been made async. This ensures that
1) requests are not blocked waiting for re-validation of tokens and
   instances
2) we dont get into situations where we flood grafana.com with
   validation requests

Instead of updating the cached entries when they are past their TTL,
We run a background tasks that scans over the cached items looking
for expired entries and re-validating them.

On reads of the cache, we update a lastRead timer.  If this timer
falls behind the last validation then we know that the cached item
is no longer being read, so we purge it from the cache.

To handle the background tasks, the AuthPlugin interface has been updated
to ensure plugins have a "Stop()" method.

Finally the auth-valid-ttl and auth-invalid-ttl cmdline args have been
replaced with a single 'auth-cache-ttl' arg, which defaults to 1hour.